Fix issue with downloading vm.tiktok.com urls
Reason is that only specific user agents are allowed are able to get
till the end of the referrer chain. Youtube-DL is getting blocked on the
way with it's default user agent

But as it turns out: `yt_dlp` has no ability to inject on an instance
basis a custom User-Agent into the requests that try to get referrers,
but as every such call is passed through the `YoutubeDL.urlopen`
function, I just rewrote this function to accept custom `http_headers`
from the `YoutubeDL` init parameters.