Prevent bot to update message to often
I can't update a message to often if it is in a group chat, I'll be rate
limited after only four updates in a short amount of time. This change
prevents that (in groups) by only showing values that are higher than a
certain bucket. This allows for easily changing the amount of update
than can be send.