Remove sharing tracking from tiktok urls
TikToks vm.tiktok.com urls are shortlinks that also contain information
about the user that shared the video which leads to TikTok suggesting on
usage of that link to follow the account that shared the video
originally.

This can be site stepped by following the referer chain and removing all
url queries from the last resulting URL (which is also the actual
permanent link to the shared post).